backend/requirements.txt
# FastAPI and dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Auth and JWT
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Supabase client (for future JWT verification)
supabase==2.3.4

# Environment variables
python-dotenv==1.0.0

# Email (placeholder for future)
fastapi-mail==1.4.1

# File uploads
aiofiles==23.2.1

# Utilities
python-dateutil==2.8.2
